refactor: rewrite demo stories in TypeScript
